Rename `micro_mechanisms` to `mechanisms`
